requirements.txt
pillow
numpy
pycryptodome
matplotlib
scipy